# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'TelU Cup Backend API'
  description: 'Dokumentasi lengkap API manajemen kompetisi internal Tel-U Cup. Mencakup autentikasi, manajemen cabang olahraga, kontingen, registrasi tim, dan bagan pertandingan gugur.'
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
paths:
...
  /api/event-photos:
    post:
      tags:
        - 'Event Photo'
      summary: 'Upload foto event untuk deteksi wajah AI'
      description: 'Panitia mengunggah foto yang diambil selama event berlangsung. Foto diupload ke Cloudinary, lalu job background dikirim ke AI Engine (FastAPI) untuk mendeteksi dan mencocokkan wajah semua peserta yang terlihat dalam foto tersebut.'
      operationId: storeEventPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              required:
                - image
              properties:
                image:
                  description: 'File foto event. Format: jpeg/png/jpg. Maksimal 5MB.'
                  type: string
                  format: binary
              type: object
      responses:
        201:
          description: 'Foto berhasil diunggah dan sedang diproses AI secara background'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Foto berhasil diunggah dan sedang diproses oleh AI.'
                  data:
                    properties:
                      id:
                        type: integer
                        example: 15
                      cloudinary_public_id:
                        type: string
                        example: telucup/event_photos/abc123
                      image_url:
                        type: string
                        example: 'https://res.cloudinary.com/demo/image/upload/v1/telucup/event_photos/abc123.jpg'
                      uploaded_by:
                        description: 'ID user (panitia) yang mengupload'
                        type: integer
                        example: 2
                      created_at:
                        type: string
                        format: date-time
                    type: object
                type: object
        422:
          description: 'Validasi gagal — bukan file gambar atau ukuran melebihi 5MB'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        403:
          description: 'Role tidak diizinkan (hanya panitia)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: 'Gagal upload ke Cloudinary atau server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  /api/my-gallery:
    get:
      tags:
        - Gallery
      summary: 'Galeri foto pemain hasil deteksi AI'
      description: 'Player melihat daftar foto event di mana wajahnya terdeteksi oleh AI. Mendukung filter berdasarkan `status` validasi.'
      operationId: myGallery
      parameters:
        -
          name: status
          in: query
          description: 'Filter berdasarkan status validasi pemain'
          required: false
          schema:
            type: string
            enum:
              - pending
              - accepted
              - rejected
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 3
                        event_photo_id:
                          type: integer
                          example: 15
                        matched_player_id:
                          type: integer
                          example: 7
                        validation_status:
                          type: string
                          example: pending
                          enum:
                            - pending
                            - accepted
                            - rejected
                        similarity_score:
                          description: 'Skor kemiripan wajah (0–1) dari model AdaFace'
                          type: number
                          format: float
                          example: 0.924
                        bounding_box:
                          description: 'Koordinat kotak wajah dalam foto'
                          properties:
                            x:
                              type: integer
                              example: 120
                            'y':
                              type: integer
                              example: 80
                            width:
                              type: integer
                              example: 64
                            height:
                              type: integer
                              example: 64
                          type: object
                          nullable: true
                        event_photo:
                          properties:
                            id:
                              type: integer
                              example: 15
                            image_url:
                              type: string
                              example: 'https://res.cloudinary.com/demo/image/upload/v1/telucup/event_photos/abc123.jpg'
                            cloudinary_public_id:
                              type: string
                              example: telucup/event_photos/abc123
                            created_at:
                              type: string
                              format: date-time
                          type: object
                        created_at:
                          type: string
                          format: date-time
                      type: object
                type: object
        404:
          description: 'Profil player belum tersedia untuk akun ini'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/my-gallery/{photo_face_id}/validate':
    patch:
      tags:
        - Gallery
      summary: 'Terima atau tolak deteksi wajah AI'
      description: 'Player mengkonfirmasi apakah wajah yang terdeteksi oleh AI benar-benar dirinya (`accepted`) atau salah deteksi (`rejected`).'
      operationId: validatePhotoFace
      parameters:
        -
          name: photo_face_id
          in: path
          description: 'ID record photo_face yang akan divalidasi'
          required: true
          schema:
            type: integer
            example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - status
              properties:
                status:
                  description: '`accepted` = wajah benar milik saya, `rejected` = bukan saya'
                  type: string
                  example: accepted
                  enum:
                    - accepted
                    - rejected
              type: object
      responses:
        200:
          description: 'Status validasi berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Status validasi berhasil diubah menjadi 'accepted'."
                  data:
                    properties:
                      id:
                        type: integer
                        example: 3
                      validation_status:
                        type: string
                        example: accepted
                      matched_player_id:
                        type: integer
                        example: 7
                      event_photo_id:
                        type: integer
                        example: 15
                      similarity_score:
                        type: number
                        format: float
                        example: 0.924
                      updated_at:
                        type: string
                        format: date-time
                    type: object
                type: object
        403:
          description: 'Anda tidak memiliki akses untuk memvalidasi foto ini'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: 'Profil player atau record photo_face tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      description: 'Format respon error standar'
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Pesan error yang menjelaskan penyebab kegagalan.'
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...